Location

+++ core/src/main/java/uk/ac/ucl/rits/inform/datasinks/emapstar/repos/IdsEffectLoggingRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 /**
  */
-public interface IdsEffectLoggingRepository extends CrudRepository<IdsEffectLogging, Integer> {
+public interface IdsEffectLoggingRepository extends CrudRepository<IdsEffectLogging, Long> {
 }
